gandi base driver in common
